rewrite music form with react hook form, fix small server bugs
